refactor: fix naming
